Limit the number of attempts to query job status to 5 retries
This fixes a bug in Telescope where it could get caught in an infinite
loop of repeatedly querying for a job and getting an unexpected result.
With this change, Telescope will make at most 5 queries if the responses
are not success or an expected failure.

This also refactors the job retrieval code into a new class,
BigQueryJobResultCollector and adds unit test coverage for that class.