Using YAPF to apply consistent code style throughout Telescope
Applies YAPF (https://github.com/google/yapf) to all source files so
that formatting is closer to the Google style guide. It's still not
totally compliant to the style guide because there's a lot that can't be
fixed mechanically, but it gets a lot of the gruntwork out of the way.