Changing travis config to use a specific YAPF version
YAPF was causing build breaks because YAPF will improve the way it
formats code from version to version, but then when YAPF changes out
from under us, we start getting build breaks unexpectedly.

This change specifies a specific version number for YAPF (0.5.0) and
updates the code to be compatible with YAPF v0.5.0 formatting.