Preventing telescope from swallowing fatal exceptions (#101)
Currently main() puts the call to generate_query within a try/except and
swallows the exception if it encounters a fatal exception. This is bad because
it causes telescope to exit with a successful error code even if it failed,
which confuses tools that rely on Telescope's exit code.

This fixes the issue by just letting the exceptions bubble out on their own.
They don't need extra logging because the exceptions themselves will describe
the failure.